Refactor: Balloon and Banner actions hpp=>h/cpp
