+++ src/openrct2/actions/BalloonPressAction.cpp
@@ -1,0 +1,48 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2020 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#include "BalloonPressAction.h"
+
+#include "GameAction.h"
+
+void BalloonPressAction::AcceptParameters(GameActionParameterVisitor& visitor)
+{
+    visitor.Visit("id", _spriteIndex);
+}
+
+void BalloonPressAction::Serialise(DataSerialiser& stream)
+{
+    GameAction::Serialise(stream);
+    stream << DS_TAG(_spriteIndex);
+}
+
+GameActions::Result::Ptr BalloonPressAction::Query() const
+{
+    auto balloon = TryGetEntity<Balloon>(_spriteIndex);
+    if (balloon == nullptr)
+    {
+        log_error("Tried getting invalid sprite for balloon: %u", _spriteIndex);
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
+    }
+    return MakeResult();
+}
+
+GameActions::Result::Ptr BalloonPressAction::Execute() const
+{
+    auto balloon = TryGetEntity<Balloon>(_spriteIndex);
+    if (balloon == nullptr)
+    {
+        log_error("Tried getting invalid sprite for balloon: %u", _spriteIndex);
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
+    }
+
+    balloon->Press();
+
+    return MakeResult();
+}
